requirements.txt
torch>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0.1
matplotlib>=3.7.0
Pillow>=10.0.0
pyproj>=3.0.0
tqdm>=4.64.0
gdal>=3.6.0